Increase polling frequency when fetching for messages
We are seeing ~1s delay in processing messages in staging/prod that come
from RabbitMQ. It seems this might be the culprit as we are introducing
a hardcoded 1s delay between polls.

This could be made configurable but 100ms is the default for kafka consumers,
so it feels like a more natural starting point.